- Refactoring
- ShipPlaceError works now correct
